tf_fuzz: Minor improvements/fixes in TF-Fuzz
Problems corrected related to failing to deallocate resources
created over the course of a test, and to incorrect accounting for
variables when using statement-level randomization (i.e., the <n>
to <m> of {} and shuffle options).  Also removed several inter-
mediate files that should not have commited in the first place.

Change-Id: I0582cbfddb70563a071fee0cb9aaf2547876d051