tf_fuzz: refactor documentation, allowing sphinx to render it
Previously, the tf_fuzz documentation was stored in a subdirectory of
the tf_fuzz folder (tf_fuzz/docs). This patch moves this documentation
to the main docs/ folder of the repository, allowing Sphinx to render
it, and making it visible in the main HTML documentation.

Other changes:

* Merge the contents of tf_fuzz.md and tf_fuzz_dir.rst into index.rst.

* Move the directory guides to a "Source Structure" subsection.

* Add cross-references to the guide.

* Add missing apt dependencies.

Throughout the docs there are references to outdated directory
structures (e.g. tools being a subdirectory of the main tf-m
repository). This has been left for a future commit, as many of these
example paths are also present in the program's help output.

Change-Id: Ia629d68ec1001ea36844faa8308b571614159279
Signed-off-by: Nik Dewally <Nik.Dewally@arm.com>