tf_fuzz: rewrite test suite generator
Rewrite the test suite generator to be compatible with the current
regression test runner.

Originally, special support for TF-Fuzz was built into the test build
system; however, this rewrite uses the out-of-tree test suite mechanism
instead. This provides all the functionality needed for TF-Fuzz suites
to run, so special support for them is no longer needed. The TF-Fuzz
related CMake flags also no longer work, so should probably be removed
in the future.

- Changes to the regression tester also broke header includes in many of
  the demo tests -- for now, these tests have been disabled.

- Test suite boilerplate is now defined with Jinja templates. This
  should make changing the layout of the generated test suite easier in
  the future.

- Only files with the .test extension in the input directory are treated
  as tests, instead of all files. The previous tester would break on
  folders with READMEs / non test content, but now these are ignored.

Change-Id: I76d282bb8f0bedff4b32cc8a367d6a1f465ceb0a
Signed-off-by: Nik Dewally <Nik.Dewally@arm.com>